Time: 6 ms (15.85%), Space: 40.2 MB (65.09%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.java
@@ -1,6 +1,6 @@
 class Solution {
     public boolean judgeSquareSum(int c) {
-        for(long a=0;a*a<=c/2;a++){
+        for(int a=0;a*a<=c/2;a++){
             double b = Math.sqrt(c-a*a);
             if(b==(int)b) return true;
         }
